Tidy up order building
If we don't use .each (.each_with_object is used here), we need to permit
what we're using from params first. We could use .reject but we'd be looping
twice and the collection from the default form could be relatively large.

Build the params hash and use it with the double-splat to pass each hash
as keyword args.